micro-text: Support Block API Version 2
