+++ blocks/src/blocks.js
@@ -31,6 +31,8 @@ import * as captionBox from './block-universal/caption-box';
 import * as labelBox from './block-universal/label-box';
 import * as tabCaptionBox from './block-universal/tab-caption-box';
 
+import * as microText from './micro/micro-text';
+
 const cocoonBlocks = [
   balloon,
   blankBox,
@@ -50,6 +52,8 @@ const cocoonBlocks = [
   captionBox,
   labelBox,
   tabCaptionBox,
+
+  microText,
 ];
 
 
@@ -117,7 +121,7 @@ import './custom/code/block.js';
 //import './block-universal/label-box/block.js';
 
 //マイクロコピー
-import './micro/micro-text/block.js';
+//import './micro/micro-text/block.js';
 import './micro/micro-balloon/block.js';
 
 //レイアウト
